% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clue_backup.R
\name{clue_backup}
\alias{clue_backup}
\title{Backup the CLUE scene}
\usage{
clue_backup(scene)
}
\arguments{
\item{scene}{\code{\link[=scene]{scene(1)}}\cr the CLUE scene to back up.}
}
\description{
Backup the CLUE scene
}
