% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clue_commodity.R
\name{clue_commodity}
\alias{clue_commodity}
\title{Set commodities and their attributes}
\usage{
clue_commodity(scene, name, demand = NULL)
}
\arguments{
\item{scene}{description}

\item{name}{\code{\link[=character]{character(1)}}\cr name of the commodity.}

\item{demand}{\code{\link[=data.frame]{data.frame(1)}}\cr table of when, where and how
much of the commodity is produced. Must have names \code{year},
\code{region} and \code{amount}.}
}
\description{
Set commodities and their attributes
}
\details{
Additional details...
}
