% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clue_conversion.R
\name{clue_conversion}
\alias{clue_conversion}
\title{Set a conversion rule}
\usage{
clue_conversion(
  scene,
  from,
  to,
  label = NULL,
  thresh = NULL,
  auto = NULL,
  mask = NULL
)
}
\arguments{
\item{scene}{description}

\item{from}{\code{\link[=character]{character(1)}}\cr a land type \emph{from} which a
conversion shall be created.}

\item{to}{\code{\link[=character]{character(1)}}\cr a land type \emph{to} which the
conversion shall be created.}

\item{label}{\code{\link[=character]{character(1)}}\cr description of the conversion
(used in the sequence plot).}

\item{thresh}{\code{\link[=integer]{integerish(1)}}\cr the number of time steps the land
type in \code{from} must remain the same before it can change into the land
type in \code{to}.}

\item{auto}{\code{\link[=integer]{integerish(1)}}\cr the number of time steps after
which the land type in \code{from} will automatically change into
\code{to}.}

\item{mask}{\code{\link[=character]{character(1)}}\cr the name of a driver (see
\code{\link{clue_driver}}) for which conversion to the specified class is
allowed.}
}
\description{
This function constructs the conversion between from one to the other land
type.
}
