% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clue_driver.R
\name{clue_driver}
\alias{clue_driver}
\title{Set drivers of land systems}
\usage{
clue_driver(scene, file, name = NULL, ...)
}
\arguments{
\item{scene}{description}

\item{file}{description}

\item{name}{description}

\item{...}{additional arguments for for writing files with
\link[terra]{writeRaster}.}
}
\description{
Set drivers of land systems
}
