% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clue_example.R
\name{clue_example}
\alias{clue_example}
\title{Build an example CLUE scene}
\usage{
clue_example(path = NULL, until = NULL, verbose = FALSE)
}
\arguments{
\item{path}{\code{\link[=character]{character(1)}}\cr The scene has by default the root
directory tempdir(), but if you want it in a particular location, specify
that in this argument.}

\item{until}{\code{\link[=character]{character(1)}}\cr The step until which the scene
shall be built, one of \code{...},}

\item{verbose}{\code{\link[=logical]{logical(1)}}\cr be verbose about building the
example scene (default \code{FALSE}).}
}
\value{
No return value, called for the side effect of creating an example
scene at the specified \code{path}.
}
\description{
This function helps setting up an example scene with custom properties.
}
\examples{
if(dev.interactive()){
# to build the full example database
clue_example(path = paste0(tempdir(), "/myScene"))

# to make the example database until a certain step
clue_example(path = paste0(tempdir(), "/myScene"), until = "")

}
}
