% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clue_initiate.R
\name{clue_initiate}
\alias{clue_initiate}
\title{Initiate a CLUE scene}
\usage{
clue_initiate(scene, module = NULL)
}
\arguments{
\item{scene}{\code{\link[=scene]{scene(1)}}\cr the CLUE scene for which to prepare the
model directory.}

\item{module}{\code{\link[=character]{character(1)}}\cr name of a module to modify a
vanilla model.}
}
\value{
no return value, called for the side-effect of creating directories
and initial files according to the model scene.
}
\description{
This creates the initial model files for a CLUE scene while ensuring in
parallel that everything is set up properly.
}
