% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clue_landsystems.R
\name{clue_landsystems}
\alias{clue_landsystems}
\title{Set the initial land systems information}
\usage{
clue_landsystems(scene, file, cats = NULL, ...)
}
\arguments{
\item{scene}{description}

\item{file}{description}

\item{cats}{(mention that either to provide this, or set it already before in
\code{map}).}

\item{...}{additional arguments for for writing files with
\link[terra]{writeRaster}.}
}
\description{
Set the initial land systems information
}
