% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clue_landtype.R
\name{clue_landtype}
\alias{clue_landtype}
\title{Set land types and their attributes}
\usage{
clue_landtype(
  scene,
  name,
  resistance = NULL,
  suitability = NULL,
  production = NULL,
  conversion = NULL,
  neighborhood = NULL,
  preference = NULL
)
}
\arguments{
\item{scene}{description}

\item{name}{\code{\link[=character]{character(1)}}\cr name of the land type.}

\item{resistance}{\code{\link[=numeric]{numeric(1)}}\cr the relative conversion
resistance to land change.}

\item{suitability}{\code{\link[=list]{list(.)}}\cr list of the result of regression
analysis. Includes the constant (\emph{beta_0}) and as many drivers and
their beta coefficient as there are (\emph{beta_1}, \emph{beta_2}, ...).
Driver names must correspond to the names defined in
\code{\link{clue_gridded}}. Each land type must have at least one
suitability driver (with \code{const = 1}).}

\item{production}{\code{\link[=list]{list(.)}}\cr optionally, list of services produced
by this land type. Each list-item is a list itself, its name the service
produced and the values of the list specify the \code{priority} and
\code{amount} of production.}

\item{conversion}{\code{\link[=list]{list(.)}}\cr optionally, list of land types into
which this land type can be converted. Each list-item is a list itself, its
name the land type into which the focal land type can be converted and the
values of the list specify the details of conversion. Must have a value for
\code{'label'} and optionally one of \code{min}, \code{max},\code{auto} or
\code{'mask'}.}

\item{neighborhood}{\code{\link[=list]{list(1)}}\cr work in progress}

\item{preference}{\code{\link[=list]{list(2)}}\cr work in progress}
}
\description{
Set land types and their attributes
}
