% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clue_mask.R
\name{clue_mask}
\alias{clue_mask}
\title{Set a mask to spatially restrict a model}
\usage{
clue_mask(scene, file)
}
\arguments{
\item{scene}{description}

\item{mask}{description}
}
\description{
Set a mask to spatially restrict a model
}
