% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clue_modify.R
\name{clue_modify}
\alias{clue_modify}
\title{Apply a modifier to a CLUE scene}
\usage{
clue_modify(scene)
}
\arguments{
\item{scene}{\code{\link[=scene]{scene(1)}}\cr the CLUE scene to mofidy.}
}
\value{
no return value, called for the side effect of ...
}
\description{
Apply a modifier to a CLUE scene
}
