% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clue_options.R
\name{clue_options}
\alias{clue_options}
\title{Set options of the current model run}
\usage{
clue_options(
  iter_mode = NULL,
  conv1 = NULL,
  conv2 = NULL,
  out_type = NULL,
  rs_reg = NULL,
  ls_hist = NULL,
  neigh = NULL,
  loc_pref = NULL,
  dyn_lusmat = NULL,
  out_write = NULL,
  iter_param = NULL,
  protected = TRUE
)
}
\arguments{
\item{iter_mode}{\code{\link[=integer]{integerish(1)}}\cr Iteration mode: \itemize{\item
0 = convergence criteria are expressed as a percentage of the demand, \item
1 = convergence criteria are expressed as absolute values, \item 2 = not
yet implemented}}

\item{conv1}{\code{\link[=numeric]{numeric(1)}}\cr First convergence criterion.}

\item{conv2}{\code{\link[=numeric]{numeric(1)}}\cr Second convergence criterion.}

\item{out_type}{\code{\link[=integer]{integerish(1)}}\cr Choice for the type of output
file:
\itemize{ \item 0 = no headers in output file, \item 1 = ArcGIS headers,
\item -2 = no headers and suppress information in log-file, \item 2 =
ArcGIS headers and suppress information in log-file, \item 3 =
ArcGIS/ArcMap extension and headers (default)}}

\item{rs_reg}{\code{\link[=integer]{integerish(1)}}\cr Region specific regressions.
\itemize{ \item 0 = no different regressions for different regions
(default), \item 1 = different regressions for different regions, \item 2
= different regressions for different regions.}}

\item{ls_hist}{\code{\link[=integer]{integerish(2)}}\cr Initialization of land use
history:
\itemize{ \item 0 = read from file age, \item 1 = generate random number +
standard seed (default), \item 2 = merely random number} For options 1 and
2, an additional value for the maximum number of years of the random
numbers must be given.}

\item{neigh}{\code{\link[=integer]{integerish(1)}}\cr Neighbourhood function (not yet
implemented): \itemize{ \item 0 = off (default), \item 1 = on in simulation
(automatically chosen on use of \code{\link{setNeighbourhood}}), \item 2 =
only influences calculated, no simulation}}

\item{loc_pref}{\code{\link[=integer]{integerish(1)}}\cr}

\item{dyn_lusmat}{\code{\link[=integer]{integerish(1)}}\cr Dynamic landuse matrix (not
yet implemented): \itemize{ \item 0 = off (default), \item 1 = on}}

\item{out_write}{\code{\link[=integer]{integerish(1)}}\cr Write selected output
measures:
\itemize{ \item 0 = off, \item 1 = on for final year (default), \item 2 =
on for all years}}

\item{iter_param}{\code{\link[=numeric]{numeric(1)}}\cr Choice for stability vs. speed
of convergence. Value should be between 0.001 and 0.1, with higher values
leading to more stable and more likely solution and lower values leading to
faster and less likely solution.}

\item{protected}{\code{\link[=logical]{logical(1)}}\cr whether or not to consider
protected areas (if the gridded object 'protected' has been given). (not
implemented yet)}
}
\value{
no return value, called for the side-effect of setting model options.
}
\description{
Set options of the current model run
}
