% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clue_run.R
\name{clue_run}
\alias{clue_run}
\title{Run the model}
\usage{
clue_run(scene, mode, stepwise)
}
\arguments{
\item{scene}{description}

\item{mode}{description}

\item{stepwise}{description}
}
\description{
Run the model
}
\details{
Additional details...
}
