% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clue_scene.R
\name{clue_scene}
\alias{clue_scene}
\title{Create a new CLUE scene}
\usage{
clue_scene(root, name, period, description = NULL)
}
\arguments{
\item{root}{\code{\link[=character]{character(1)}}\cr path where the model files will be
stored.}

\item{name}{\code{\link[=character]{character(1)}}\cr the name of the scene.}

\item{period}{\code{\link[=integer]{integerish(2)}}\cr start and end year of the scene.}

\item{description}{\code{\link[=character]{character(1)}}\cr the description of the
scene.}
}
\value{
an empty CLUE scene that captures initial metadata of the CLUE model
run, but doesn't contain any parameters yet.
}
\description{
Create a new CLUE scene
}
