% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clue_validate.R
\name{clue_validate}
\alias{clue_validate}
\title{Validate the current status of a model}
\usage{
clue_validate(scene, verbose = FALSE)
}
\arguments{
\item{scene}{\code{\link[=scene]{scene(1)}}\cr the CLUE scene to validate.}

\item{verbose}{\code{\link[=logical]{logical(1)}}\cr whether or not to give the full
diagnosis output, or just until the most recent error.}
}
\value{
no return value, called for the side-effect of checking the current
setup.
}
\description{
This checks all directories and files and gives helpful feedback.
}
