% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clueRIO-package.R
\docType{package}
\name{clueRIO-package}
\alias{clueRIO}
\alias{clueRIO-package}
\title{clueRIO: Input/Output management for CLUMondo}
\description{
The
\href{https://www.environmentalgeography.nl/site/data-models/models/clumondo-model/}{CLUMondo}
model is a dynamic, spatially explicit, land use and land cover change model.
clueRIO manages input/output for setting up the file-structure required for
successfully configuring and running a CLUE model. This acts as an interface
that enables users to build and document scenarios based on such models
quickly and in a script-based fashion.
}
\details{
Please check the above link for more information on the CLUMondo
model. The most up-to-date version is typically a command line version.
This R-package has the aim to simplify specifying all files needed to
successfully run that version. Please make sure to reference not only this
R-package, but also the references on the website and in particular
\emph{"Van Asselen, S., Verburg, P.H., 2013. Land cover change or land-use
intensification: simulating land system change with a global-scale land
change model. Global Change Biology 19, 3648–3667"}.
}
\seealso{
\itemize{ \item Github project:
\href{https://github.com/luckinet/cluerio}{https://github.com/luckinet/cluerio}
\item Report bugs:
\href{https://github.com/luckinet/cluerio/issues}{https://github.com/luckinet/cluerio/issues}
}
}
\author{
\strong{Maintainer, Author}: Steffen Ehrmann
\email{steffen.ehrmann@posteo.de}
}
