% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeModel.R
\name{makeModel}
\alias{makeModel}
\title{Build all files required for the model}
\usage{
makeModel(stepwise = FALSE)
}
\value{
No return value, called for the side effect of creating all the basic
files needed to run CLUMondo (such as \emph{cov_all.0.asc},
\emph{demand.in1}, \emph{allc1.reg} or \emph{main.1}, etc) in the project
directory.
}
\description{
This function generates the tables that are needed to run CLUMondo.
}
