% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{mod_crop_rotation}
\alias{mod_crop_rotation}
\title{Crop rotation modifier}
\usage{
mod_crop_rotation()
}
\description{
Crop rotation modifier
}
\details{
To implement crop rotations, \code{'locspec'} files are adapted for
each year. The yearly \code{'locspec'} is calculated by \enumerate{\item
multiplying the \code{'age'} layer with the \code{'locspec weight factor'}
and \item selecting only a subset of pixels, depending on the sign of this
weight factor. In case the sign is negative, it would keep those pixels
that were covered by the focal class in the previous time-step and thus
down-weigh where is was present previously. In case the sign is positive,
it would keep those pixels that were covered by all other classes in the
previous time-step and thus up-weigh where it was not present previously.}
}
