% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regCommodity.R
\name{regCommodity}
\alias{regCommodity}
\title{Set commodities}
\usage{
regCommodity(name = NULL, match = "exact", table = NULL)
}
\arguments{
\item{name}{\code{\link[=character]{character(1)}}\cr the name of the commodity.}

\item{match}{\code{\link[=character]{character(1)}}\cr the type of match for this commodity,
either \code{"exact"} (the default), \code{"minimum"} (the demand indicated
is the minimum demand that needs to be achieved, any overshoot is accepted)
or \code{"maximum"} (the demand indicated is the maximum that may be
reached, any undershoot is accepted).}
}
\description{
This function registers commodities and associates them to land-use types.
}
\seealso{
Other functions to set new items: 
\code{\link{regLanduse}()}
}
\concept{functions to set new items}
