% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regDemand.R
\name{regDemand}
\alias{regDemand}
\title{Set the table of demand}
\usage{
regDemand(table = NULL, values = NULL, region = NULL)
}
\arguments{
\item{table}{\code{\link[=data.frame]{data.frame(1)}}\cr}

\item{region}{\code{\link[=character]{character(1)}}\cr the column/variable in
\code{table} that distinguishes unique regions of that object.}
}
\description{
This function
}
\seealso{
Other functions to register items: 
\code{\link{regGridded}()},
\code{\link{regRegions}()}
}
\concept{functions to register items}
