% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regGridded.R
\name{regGridded}
\alias{regGridded}
\title{Set gridded objects}
\usage{
regGridded(
  name = NULL,
  dir = NULL,
  pattern = NULL,
  crs = NULL,
  obj = NULL,
  table = NULL,
  ...
)
}
\arguments{
\item{name}{\code{\link[=character]{character(1)}}\cr the name of the driving factor.}

\item{dir}{\code{\link[=character]{character(1)}}\cr path of the directory where to find the
layers of this driver.}

\item{pattern}{\code{\link[=character]{character(1)}}\cr a regular expression to detect all
files that are part of that driving factor.}
}
\description{
This function makes an inventory of driving factors by recording the name and
exact path of each gridded layer that shall be used as driving factor.
}
\details{
If there is more than one layer selected via \code{pattern}, the
function makes sure that for this driver there is a layer for each year
that has been defined by \code{\link{newProject}}.

\code{name = "landuse"} is reserved to register the mandatory intial
landuse layer and \code{name = "region"} for the optional area under
consideration. \code{"landuse"} must be provided with an attribute table
that contains all the landuse systems that shall be modelled (see
\code{\link{setLAT}}).
}
\seealso{
Other functions to register items: 
\code{\link{regDemand}()},
\code{\link{regRegions}()}
}
\concept{functions to register items}
