% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regLanduse.R
\name{regLanduse}
\alias{regLanduse}
\title{Set land-use types}
\usage{
regLanduse(name = NULL, resistance = NULL, limit = NULL, table = NULL)
}
\arguments{
\item{name}{\code{\link[=character]{character(1)}}\cr the name of the land-use type.}

\item{resistance}{\code{\link[=numeric]{numeric(1)}}\cr the relative conversion
resistance to land-use change.}

\item{limit}{\code{\link[=integer]{integerish(1)}}\cr the number of time-steps a
land-use type is allowed to remain in a location at most.}

\item{table}{\code{\link[=data.frame]{data.frame(1)}}\cr table of information on
land-use types. Must contain the columns \code{name}, \code{resistance},
\code{limit} and \code{target}.}
}
\description{
This function registers land-use types.
}
\seealso{
Other functions to set new items: 
\code{\link{regCommodity}()}
}
\concept{functions to set new items}
