% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regRegions.R
\name{regRegions}
\alias{regRegions}
\title{Set distinct regions to model}
\usage{
regRegions(features = NULL, region = NULL)
}
\arguments{
\item{features}{\code{\link[=sf]{sf(.)}}\cr collection of simple features that outline
the distinct regions for each of which an iteration of the model shall be
run.}

\item{region}{\code{\link[=character]{character(1)}}\cr the column/variable in
\code{features} that distinguishes unique regions of that object. The
values in this column need to correspond to \code{region} in
\code{\link{regDemand}}.}
}
\description{
This function registers the spatial outline of one or more regions.
}
\details{
The CLUMondo model is currently meant to be run for a distinct
region only. A project with several regions (e.g., territorial units) is
basically split up into several runs of the model, each for a specific
region.
}
\seealso{
Other functions to register items: 
\code{\link{regDemand}()},
\code{\link{regGridded}()}
}
\concept{functions to register items}
