% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel.R
\name{runModel}
\alias{runModel}
\title{Run the model}
\usage{
runModel(region = NULL, mode = NULL, stepwise = FALSE, backup = FALSE)
}
\arguments{
\item{mode}{\code{\link[=integer]{integerish(1)}}\cr the model mode.}

\item{backup}{\code{\link[=logical]{logical(1)}}\cr whether or not the output of the
previous run shall be stored in a folder where it is not overwritten.}
}
\value{
No return value, called for the side effect of preparing a CLUMondo
model and running it.
}
\description{
Run the model
}
\details{
This function creates a meta-data object in the main directory,
calls \code{\link{makeInput}} to build all the input files required by
CLUMondo and sends a call to the OS console to run CLUMondo first to create
probability maps, then to allocate.
}
