% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1scene.R
\docType{class}
\name{scene-class}
\alias{scene-class}
\alias{scene}
\title{CLUE scene class (S4) and methods}
\description{
A \code{scene} stores metadata and configuration of a CLUE model.
}
\section{Slots}{

\describe{
\item{\code{meta}}{\code{\link[=list]{list(3)}}\cr}

\item{\code{period}}{\code{\link[=numeric]{numeric(2)}}\cr}

\item{\code{landtypes}}{\code{\link[=list]{list(.)}}\cr}

\item{\code{grids}}{\code{\link[=list]{list(.)}}\cr}

\item{\code{goods}}{\code{\link[=list]{list(.)}}\cr}

\item{\code{validated}}{\code{\link[=logical]{logical(1)}}\cr}
}}

