% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setConversion.R
\name{setConversion}
\alias{setConversion}
\title{Set a conversion rule}
\usage{
setConversion(
  from = NULL,
  to = NULL,
  label = NULL,
  thresh = NULL,
  auto = NULL,
  mask = NULL
)
}
\arguments{
\item{from}{\code{\link[=character]{character(1)}}\cr a landuse type for which a
relationship shall be created.}

\item{to}{\code{\link[=character]{character(1)}}\cr a landuse type to which the
relationship shall be created.}

\item{label}{\code{\link[=character]{character(1)}}\cr description of the relationship.}

\item{thresh}{\code{\link[=integer]{integerish(1)}}\cr the number of time-steps the
landuse type in \code{from} must remain the same before it can change into
the landuse type in \code{to}.}

\item{auto}{\code{\link[=integer]{integerish(1)}}\cr the number of time-steps after
which the landuse type in \code{from} will automatically change into
\code{to}.}

\item{mask}{\code{\link[=character]{character(1)}}\cr the name of a gridded object for
which conversion to the specified class is allowed.}
}
\description{
This function constructs the conversion between two landuse types or a
commodity and a landuse type.
}
\details{
From the resulting land-use change sequence, a land-use conversion
matrix is derived.
}
\seealso{
Other functions to specify landuse: 
\code{\link{setNeighbourhood}()},
\code{\link{setPreference}()},
\code{\link{setSuitability}()}
}
\concept{functions to specify landuse}
