% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setLAT.R
\name{setLAT}
\alias{setLAT}
\title{Set a raster attribute table of the landuse grid (LAT)}
\usage{
setLAT(table = NULL)
}
\arguments{
\item{table}{\code{\link[=data.frame]{data.frame(1)}}\cr table that contains the columns
\code{'system'} and \code{'ID'} of the unique values in the initial landuse
map.}
}
\description{
Land-use attribute table
}
\concept{functions to specify gridded objects}
