% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setNeighbourhood.R
\name{setNeighbourhood}
\alias{setNeighbourhood}
\title{Specify neighbourhood models}
\usage{
setNeighbourhood(
  landuse = NULL,
  const = NULL,
  ...,
  weight = NULL,
  kernel = matrix(1, 3, 3),
  table = NULL
)
}
\arguments{
\item{landuse}{\code{\link[=character]{character(1)}}\cr the landuse type for which to
specify the neighbourhood parameters.}

\item{const}{\code{\link[=numeric]{numeric(1)}}\cr constant of regression equation (how
neighborhood results contribute to the suitability of a location) for this
landuse type.}

\item{...}{\code{\link[=numeric]{numeric(1)}}\cr named value, where the name is the
landuse type in the neighbourhood and the value is the respective beta
coefficient of the regression equation (how neighborhood results contribute
to the suitability of a location).}

\item{weight}{\code{\link[=numeric]{numeric(1)}}\cr the weight (0 - 1) assigned to the
neighborhood function.}

\item{kernel}{\code{\link[=matrix]{matrix(1)}}\cr the region within which neighbourhood
shall be considered. This is a matrix with an odd number of columns and
rows containing weights (0 - 1) of the cells in the neighbourhood.}

\item{table}{\code{\link[=data.frame]{data.frame(1)}}\cr object that contains one or
more landuse types in rows and the columns \code{"landuse"} and
\code{"const"} of each type and any number of landuse types with the
respective model coefficients. (not yet implemented)}
}
\description{
This function sets the parameters for neigbourhood modelling. This requires
to specify the neighbourhood (dimensions and weights) and a regression
equation that determines how neighborhood results contribute to the
suitability of a location.
}
\details{
When a land use type is not influenced by a neighborhood function
(while others are), \code{const} is set to 0 and nothing else is specified.
}
\seealso{
Other functions to specify landuse: 
\code{\link{setConversion}()},
\code{\link{setPreference}()},
\code{\link{setSuitability}()}
}
\concept{functions to specify landuse}
