% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPreference.R
\name{setPreference}
\alias{setPreference}
\title{Set a location specific preference}
\usage{
setPreference(landuse = NULL, layer = NULL, weight = NULL, table = NULL)
}
\arguments{
\item{landuse}{\code{\link[=character]{character(1)}}\cr the landuse type for which to
specify location specific preference.}

\item{layer}{\code{\link[=character]{character(1)}}\cr the gridded object in which the
location specific preference values are noted. If this is left empty, it
will be a layer that includes only the value 0 for all pixels.}

\item{weight}{\code{\link[=numeric]{numeric(1)}}\cr the weight (beween -1 for
down-weighing and 1 for up-weighing the preferences) assigned to this
factor.}

\item{table}{\code{\link[=data.frame]{data.frame(1)}}\cr table of information on
location specific preference. Must contain the columns \code{landuse},
\code{layer}, and \code{weight}.}

\item{subset}{\code{\link[=character]{character(1)}}\cr an expression that successfully
evaluates to the subset of a raster, for example to subset a layer on the
fly between years.}
}
\description{
Set a location specific preference
}
\details{
\code{layer = "self"} is reserved to derive dynamic locspec maps. In
that case the age map of each timestep is multiplied by \code{weight}.
Depending on the sign of weight, either all age-value of this class
(negative) weight) or of all other classes (positive weight) are selected
for multiplication (all others are set to 0). De facto, this leads to
decrease of preference where the class already occurs with a negative sign,
and increase of preference where the class does not yet exist with a
positive sign. This can be interessting, when a class becomes less likely,
the longer it is present, e.g., in crop rotation schemes, where both,
ressource depletion or pathogen pressure but also increasing times of
commodities that repair the soil, make rotation more likely.
}
\seealso{
Other functions to specify landuse: 
\code{\link{setConversion}()},
\code{\link{setNeighbourhood}()},
\code{\link{setSuitability}()}
}
\concept{functions to specify landuse}
