% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setProduction.R
\name{setProduction}
\alias{setProduction}
\title{Set production of commodities by land-use types}
\usage{
setProduction(
  of = NULL,
  by = NULL,
  amount = NULL,
  priority = NULL,
  exclude = NULL
)
}
\arguments{
\item{of}{\code{\link[=character]{character(1)}}\cr a commoditiy for which a production
amount should be set.}

\item{by}{\code{\link[=character]{character(.)}}\cr land-use type(s) by which \code{of}
is produced.}

\item{amount}{\code{\link[=numeric]{numeric(.)}}\cr the amount(s) of produced (must have
same length as \code{by}).}

\item{priority}{\code{\link[=integer]{integerish(1)}}\cr how likely the commodity is
produced by the land-use type(s) in response to other land-use types.}

\item{exclude}{\code{\link[=character]{character(.)}}\cr land-use types from which the
commodity is excluded to be produced. Overrides the other arguments if set.}

\item{unit}{\code{\link[=character]{character(1)}}\cr the unit of \code{amount}.}
}
\description{
This function registers how much of a good or service (commodity) is produced
by land-use type(s).
}
\details{
A commodity can be any good or service and it can be produced by
several land-use types. However, often distinct land-use types produce
different amounts of the commodities and they produce them with a different
likeliness. This likeliness is specified as land-use type specific
\code{priority} value.
}
\concept{functions to specify commodities}
