% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setSuitability.R
\name{setSuitability}
\alias{setSuitability}
\title{Specify suitability models}
\usage{
setSuitability(landuse = NULL, const = NULL, ..., table = NULL)
}
\arguments{
\item{landuse}{\code{\link[=character]{character(1)}}\cr the landuse type for which to
specify suitability parameters.}

\item{const}{\code{\link[=numeric]{numeric(1)}}\cr constant of regression equation for
this landuse type.}

\item{...}{named value, where the name is the
gridded object name and the value is the beta coefficient of the regression
equation.}

\item{table}{\code{\link[=data.frame]{data.frame(1)}}\cr object that contains one or
more land-use types in rows and the columns \code{"landuse"} and
\code{"const"} of each type and any number of explanatory factors with the
respective model coefficients. Explanatory factor names must correspond to
the drivers set in \code{\link{setDriver}}}
}
\description{
This function sets the parameters for suitability modelling.
}
\seealso{
Other functions to specify landuse: 
\code{\link{setConversion}()},
\code{\link{setNeighbourhood}()},
\code{\link{setPreference}()}
}
\concept{functions to specify landuse}
