% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1scene.R
\name{show,scene-method}
\alias{show,scene-method}
\title{Print scene in the console}
\usage{
\S4method{show}{scene}(object)
}
\arguments{
\item{object}{\code{\link[=scene]{scene(1)}}\cr object to \code{show}.}
}
\description{
Print scene in the console
}
