% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showAnimation.R
\name{showAnimation}
\alias{showAnimation}
\title{Animate landuse change}
\usage{
showAnimation(interval = 1, input = FALSE, target = "cover")
}
\arguments{
\item{interval}{\code{\link[=numeric]{numeric(1)}}\cr time increment of the animation.}

\item{input}{\code{logical(1)}][logical]\cr whether or not to also include the
gridded objects (sc1gr*.fil) in the animation.}
}
\description{
Animate landuse change
}
