% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showSequence.R
\name{showSequence}
\alias{showSequence}
\title{Graph of landuse conversion sequence}
\usage{
showSequence(self = FALSE)
}
\arguments{
\item{self}{\code{\link[=logical]{logical(1)}}\cr whether or not to show edges for when
landuse types can be converted to themselves.}
}
\description{
Graph of landuse conversion sequence
}
