% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showSuitability.R
\name{showSuitability}
\alias{showSuitability}
\title{Plot of suitabilities}
\usage{
showSuitability(target = NULL, comparison = NULL, zlim = c(0, 1), steps = 100)
}
\arguments{
\item{target}{\code{\link[=character]{character(1)}}\cr the focal driver.}

\item{comparison}{\code{\link[=character]{character(.)}}\cr the drivers to compare the
focal driver with.}

\item{zlim}{\code{\link[=numeric]{numeric(2)}}\cr}
}
\description{
This function shows suitability maps by focusing on a 'target' driver and the
combination of all 'comparison' drivers.
}
