% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startProject.R
\name{startProject}
\alias{startProject}
\title{Start a new project}
\usage{
startProject(
  root = NULL,
  name = NULL,
  start = NULL,
  end = NULL,
  landuse = NULL,
  demands = NULL,
  resolution = NULL,
  extent = NULL
)
}
\arguments{
\item{root}{\code{\link[=character]{character(1)}}\cr path to the root directory that contains
or shall contain the working directory of this project.}

\item{name}{\code{\link[=character]{character(1)}}\cr the name for this project.}

\item{start}{\code{\link[=integer]{integerish(1)}}\cr the first year of the project.}

\item{end}{\code{\link[=integer]{integerish(1)}}\cr the last year of the project.}

\item{landuse}{\code{\link[=character]{character(.)}}\cr the allowed/valid landuse systems for
this project.}

\item{demands}{\code{\link[=character]{character(.)}}\cr the allowed/valid demand types for
this project.}

\item{resolution}{\code{\link[=numeric]{numeric(2)}}\cr vector of x and y resolution.}

\item{extent}{\link{\code{extent(1)}}\cr maximum extent of the region,
raster::extent object.}
}
\value{
Enters the provided information as 'lclu' into the options.
}
\description{
This function registers initial metadata of the current project into the
global options.
}
