% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useModifier.R
\name{useModifier}
\alias{useModifier}
\title{Modify parts of a model}
\usage{
useModifier(module = "crop_rotation", ...)
}
\arguments{
\item{module}{\code{\link[=character]{character(1)}}\cr modifications of CLUMondo
typically come with a set of files that have to be adapted and these are
summarised into modules. By selecting a module you thus make sure that all
adaptions are done consistently.}

\item{...}{module specific objects for that modification. See documentation
of the respective module.}
}
\description{
Modify parts of a model
}
\details{
Availble modules \itemize{ \item \code{\link{mod_crop_rotation}}:
When using this module, CLUMondo will be run separately for each year by
default and re-calculate \code{'locspec'} after each step.}
}
