% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clue_gridded.R
\name{clue_gridded}
\alias{clue_gridded}
\title{Set drivers of land systems}
\usage{
clue_gridded(scene, file, type, name = NULL, ...)
}
\arguments{
\item{scene}{description}

\item{file}{description}

\item{type}{description}

\item{name}{description}

\item{...}{additional arguments for for writing files with
\link[terra]{writeRaster}.}
}
\description{
Set drivers of land systems
}
