% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clue_initiate.R
\name{clue_initiate}
\alias{clue_initiate}
\title{Prepare the CLUE model directory}
\usage{
clue_initiate(scene, clue.exe = NULL)
}
\arguments{
\item{scene}{\code{\link[=scene]{scene(1)}}\cr the CLUE scene for which to prepare the
model directory.}

\item{clue.exe}{\code{\link[=character]{character(1)}}\cr by default, this package ships
with the publicly available version of CLUMondo. Optionally, you can
specify here the full path to another \emph{clue.exe} file with which the
model should be run.}
}
\value{
no return value, called for the side-effect of creating directories
and initial files.
}
\description{
This creates directories and initial files from a CLUE scene.
}
\details{
Additional details...
}
