% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clue_landsystems.R
\name{clue_landsystems}
\alias{clue_landsystems}
\title{Set the initial land systems information}
\usage{
clue_landsystems(scene, file, regions, cats = NULL)
}
\arguments{
\item{scene}{description}

\item{file}{description}

\item{regions}{description}

\item{cats}{description}
}
\description{
Set the initial land systems information
}
\details{
Additional details...
}
