% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clue_landtype.R
\name{clue_landtype}
\alias{clue_landtype}
\title{Set land types and their attributes}
\usage{
clue_landtype(
  scene,
  name,
  resistance,
  suitability,
  production = NULL,
  conversion = NULL,
  neighborhood = NULL,
  preference = NULL
)
}
\arguments{
\item{scene}{\code{\link[=scene]{scene(1)}}\cr the CLUE scene to which to add a land
type.}

\item{name}{\code{\link[=character]{character(1)}}\cr name of the land type.}

\item{resistance}{\code{\link[=numeric]{numeric(1)}}\cr the relative conversion
resistance to land change.}

\item{suitability}{\code{\link{data.frame}}\cr \emph{obligatory} table of
the result of regression analysis. Has at least two rows. Must have columns
\itemize{
\item \code{driver}: from line 2 onward the name of the driver (must
correspond to names of gridded items of \code{type = "driver"} as
defined in \code{\link{clue_gridded}}) and
\item \code{beta}: the respective regression coefficient
(beta_\{1, 2, ...\}).}
The first row must have the driver \code{"const"} with the respective
\emph{beta_0} value.}

\item{production}{\code{\link{data.frame}}\cr optional table of goods and
services produced by this land type (for the demand to produce, see
\code{\link{clue_goods}}). Each row contains one good. Must have 3 columns
\itemize{
\item \code{name}: the name of the good or service,
\item \code{amount}: how much each cell in the map produces and
\item \code{priority}: an integer value that indicates the priority with
which the good or service is produced by the focal land type.
If a good is produced by more than one land type, the land type with
a smaller number here is produced with a higher priority.}}

\item{conversion}{\code{\link{data.frame}}\cr optional table of land types
into which this land type can be converted. Each row contains one type into
which it can be converted. Must have a columns \itemize{
\item \code{to}: name of the land type into which to convert and
\item \code{label}: name of the conversion process,
} and optionally, for the details of conversion one of \itemize{
\item \code{min}: number of time-steps that need to pass until the
conversion can take place,
\item \code{max}: number of time-steps within which the conversion can
take place,
\item \code{auto}: number of time-steps after which a conversion must take
place (from local dynamics and not to fulfill demand of a good or
service) or
\item \code{mask}: name of a gridded item (with values 0, 1 and NA/-9999)
that indicates with value 1 where conversion can take place and with
0, where they cannot take place.}}

\item{neighborhood}{\code{\link{data.frame}}\cr optional table of
neighborhood parameters. Has at least four rows. Must have columns
\itemize{
\item \code{type}: from line 4 onward the name of the land type that
influences the suitability of the focal land type if it's in the
neighborhood (must correspond to names of other land types defined
with this function) and
\item \code{beta}: the respective regression coefficient
(beta_\{1, 2, ...\})}
The first row has the number of a kernel that must have been defined with
\code{clue_gridded} and the second row the weight assigned to the
neighborhood function, and the third row must have the type \code{"const"}
with the respective \emph{beta_0} value.}

\item{preference}{\code{\link{data.frame}}\cr optional table of suitability
preferences. One row only. Must have columns \code{layer} and
\code{weight}. Layer names must correspond to names of gridded items of
\code{type = "preference"} as defined in \code{\link{clue_gridded}}.}
}
\description{
Set land types and their attributes
}
\section{Suitability}{


mention specifics about suitabilities
}

\section{Production}{


mention specifics about production
}

\section{Conversion}{


Conversions are implicitly always set from a focal class to itself. By
default, this conversion is allowed indefinitely and one does not have to
set it. However, if one wants to limit the persistence of a land type to,
e.g., 5 years in a row, one needs to specify

\if{html}{\out{<div class="sourceCode">}}\preformatted{clue_landtype(...,
              name = "it",
              conversion = tibble(to = "it",
                                  label = "limit_self",
                                  max = 5))
}\if{html}{\out{</div>}}
}

\section{Neighborhood}{


mention specifics about neighborhood
}

\section{Preference}{


mention specifics about preferences
}

