% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clue_landtype.R
\name{clue_landtype}
\alias{clue_landtype}
\title{Set land types and their attributes}
\usage{
clue_landtype(
  scene,
  name,
  resistance = NULL,
  suitability = NULL,
  production = NULL,
  conversion = NULL,
  neighborhood = NULL,
  preference = NULL
)
}
\arguments{
\item{scene}{description}

\item{name}{\code{\link[=character]{character(1)}}\cr name of the land type.}

\item{resistance}{\code{\link[=numeric]{numeric(1)}}\cr the relative conversion
resistance to land change.}

\item{suitability}{\code{\link[=data.frame]{data.frame(2)}}\cr obligatory table of the
result of regression analysis. At least two rows. Includes the constant
(\emph{beta_0}) and as many drivers and their beta coefficient as there are
in the regression formula (\emph{beta_1}, \emph{beta_2}, ...). Must have
the column \code{const} and at least one other column with the name of the
respective suitability driver. Driver names must correspond to names of
gridded items of \code{type = "driver"} as defined in
\code{\link{clue_gridded}}.}

\item{production}{\code{\link[=data.frame]{data.frame(3)}}\cr optional table of goods
and services produced by this land type. Each row contains one good. Must
have column \code{goods} for its name, \code{amount} of production and
\code{priority}.}

\item{conversion}{\code{\link[=data.frame]{data.frame(.)}}\cr optional table of land
types into which this land type can be converted. Each row contains one
type into which it can be converted. Must have a column \code{to} and
\code{label} and for the details of conversion, optionally one of
\code{min}, \code{max},\code{auto} or \code{mask}.}

\item{neighborhood}{\code{\link[=data.frame]{data.frame(.)}}\cr work in progress}

\item{preference}{\code{\link[=data.frame]{data.frame(2)}}\cr optional table of
suitability preferences. One row only. Must have columns \code{layer} and
\code{weight}. Layer names must correspond to names of gridded items of
\code{type = "preference"} as defined in \code{\link{clue_gridded}}.}
}
\description{
Set land types and their attributes
}
