% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clue_landtype.R
\name{clue_landtype}
\alias{clue_landtype}
\title{Set land types and their attributes}
\usage{
clue_landtype(
  scene,
  name,
  resistance = NULL,
  limit = NULL,
  suitability = NULL,
  production = NULL,
  conversion = NULL,
  neighborhood = NULL,
  preference = NULL
)
}
\arguments{
\item{scene}{description}

\item{name}{\code{\link[=character]{character(1)}}\cr name of the land type.}

\item{resistance}{\code{\link[=numeric]{numeric(1)}}\cr the relative conversion
resistance to land change.}

\item{limit}{\code{\link[=integer]{integerish(1)}}\cr the number of time steps a
land-use type is allowed to remain in a location at most.}

\item{suitability}{\code{\link[=data.frame]{data.frame(2)}}\cr table of driver and
coefficient of suitability. Must have names \code{'driver'} and
\code{'coef'}. Driver names must correspond to the names defined in
\code{\link{clue_driver}}.}

\item{production}{\code{\link[=data.frame]{data.frame(3)}}\cr table of what, with which
priority and how much this land type produces. Must have names
\code{'commodity'}, \code{'priority'} and \code{'amount'}.}

\item{conversion}{\code{\link[=data.frame]{data.frame(.)}}\cr table of land types into
which the land type can be converted. Must have names \code{'to'} and
\code{'label'} and optionally one of \code{'thresh'}, \code{'auto'} or
\code{'mask'}.}

\item{neighborhood}{\code{\link[=data.frame]{data.frame(1)}}\cr work in progress}

\item{preference}{\code{\link[=data.frame]{data.frame(2)}}\cr table of where and how
location specific preferences are used. Must have names \code{'layer'} and
\code{'weight'}.}
}
\description{
Set land types and their attributes
}
