% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clue_landtype.R
\name{clue_landtype}
\alias{clue_landtype}
\title{Set land types and their attributes}
\usage{
clue_landtype(
  scene,
  name,
  resistance = NULL,
  limit = NULL,
  suitability = NULL,
  production = NULL
)
}
\arguments{
\item{scene}{description}

\item{name}{\code{\link[=character]{character(1)}}\cr name of the land type.}

\item{suitability}{\code{\link[=data.frame]{data.frame(1)}}\cr table of driver and
coefficient of suitability. Must have names \code{driver} and \code{coef}.
Driver names must correspond to the names defined in
\code{\link{clue_driver}}.}

\item{production}{\code{\link[=data.frame]{data.frame(1)}}\cr table of what, with which
priority and how much this land type produces. Must have names
\code{commodity}, \code{priority} and \code{amount}.}
}
\description{
Set land types and their attributes
}
\details{
Additional details...
}
