% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clue_run.R
\name{clue_run}
\alias{clue_run}
\title{Run the model}
\usage{
clue_run(scene, clue.exe, mode)
}
\arguments{
\item{scene}{description}

\item{clue.exe}{\code{\link[=character]{character(1)}}\cr R packages cannot contain
executable files, such as the CLUMondo main model executable. Therefore,
you need to download this
\href{https://www.environmentalgeography.nl/site/data-models/models/clumondo-model/}{CLUMondo*.exe}
from the official model website and provide the path to your local copy
here.}

\item{mode}{\code{\link[=integer]{integerish(1)}}\cr the model mode.}
}
\value{
no return value, called for the side-effect of calling \code{scene}
and running the CLUMondo model.
}
\description{
Run the model
}
\details{
Additional details...
}
