% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clue_scene.R
\name{clue_scene}
\alias{clue_scene}
\title{Create a new CLUE scene}
\usage{
clue_scene(root, name, period, regions, description = NULL)
}
\arguments{
\item{root}{\code{\link[=character]{character(1)}}\cr path where the model files will be
stored.}

\item{name}{\code{\link[=character]{character(1)}}\cr the name of the scene.}

\item{period}{\code{\link[=integer]{integerish(2)}}\cr start and end year of the scene.}

\item{regions}{description}

\item{description}{\code{\link[=character]{character(1)}}\cr the description of the
scene.}
}
\value{
object of class \code{\link{scene}} that contains initial meta data.
}
\description{
Create a new CLUE scene
}
\examples{
library(sf)
region <- st_read(
  paste0(system.file("test_data", package = "clueRIO"),
         "/test_region.gpkg"))

minimal <- clue_scene(
    root = paste0(tempdir(), "/ls2d1"),
    name = "ls2d1",
    period = c(2000, 2024),
    regions = region,
    description = "this is a minimal showcase where forest is converted to
                   cropland to produce crops.")
}
